requirements.txt
aiohttp>=3.8.0
requests>=2.28.0
python-dotenv>=1.0.0
eth-account>=0.8.0
lighter>=1.0.0
x10-perpetual-sdk>=1.0.0
websockets>=10.4
